Fix WebNN test only builds due to missing DirectML.dll.
Building the WebNN E2E test target will fail to run because of the missing DML binary in the output build directory that is only copied when the whole project gets built. To fix, the DirectML.dll dependency is specified as a data_deps for the webnn_native_sources build target.